requirements.txt
openai==1.56.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1